Remove All Unused Scripts + More Consistent Script Naming (#765)
* Remove Unusued Scripts

```rb
scripts = `find ./scripts -type f -maxdepth 1`.chomp.split("\n")

unused_files = []

for script in scripts
    basename = File.basename(script)
    STDERR.puts "Searching for #{basename}..."
    ag_files = `ag --count #{basename}`.chomp.split("\n")
    if ag_files.count > 0
        STDERR.puts "Found #{basename} in #{ag_files.count} files:"
        for file in ag_files
            STDERR.puts "\t-#{file}"
        end
    else
        STDERR.puts "Could not find #{basename} in any files!"
        unused_files << script
    end
end

STDERR.puts "Unused files: #{unused_files}"
puts "#{unused_files.map { |f| "'#{f}'" }.join(" ")}"
```

* Rename all `camelCase` scripts to `snake_case` for consistency (except mergeLef.py)

```rb
scripts = `find ./scripts -type f -maxdepth 1`.chomp.split("\n")

unused_files = []

for script in scripts
    if script.downcase != script
        puts script
    end
end
```

* More naming consistency, folded useless scripts
