+++ scripts/fake_diode_replace.py
